+++ App/Sources/Feature/GoalReadingFeature/Component/SettingGoalReadingView.swift
@@ -1,0 +1,19 @@
+//
+//  SettingGoalReadingView.swift
+//  MindWay
+//
+//  Created by 이승화 on 4/12/24.
+//  Copyright © 2024 team.mindway. All rights reserved.
+//
+
+import SwiftUI
+
+struct SettingGoalReadingView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    SettingGoalReadingView()
+}
