+++ App/Sources/Feature/LoginFeature/LoginView.swift
@@ -12,7 +12,7 @@ struct LoginView: View {
                 HStack(spacing: 8) {
                     Spacer()
                     
-                    MindWayAsset.Images.gAuthLogo.swiftUIImage
+                    MindWayAsset.Images.gauthLogo.swiftUIImage
                     
                     MindWayAsset.Images.signIn.swiftUIImage
