+++ App/Sources/Feature/MainFeature/Component/GoalReadingGraph.swift
@@ -59,3 +59,8 @@ struct GoalReadingGraph: View {
         }
     }
 }
+
+
+#Preview {
+    SettingGoalReadingView()
+}
